Rename to block-reward, removed TvlProvider trait
