+++ Cargo.toml
@@ -3,7 +3,6 @@ members = [
     "frame/block-reward",
     "frame/custom-signatures",
     "frame/dapps-staking",
-    "frame/reward-distribution",
     "precompiles/dapps-staking",
     "precompiles/sr25519",
     "precompiles/utils",
