--- frame/reward-distribution/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "pallet-reward-distribution"
-version = "1.0.0"
-authors = ["Stake  Technologies <devops@stake.co.jp>"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://astar.network"
-repository = "https://github.com/AstarNetwork/Astar"
-description = "FRAME pallet for managing reward distribution"
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.2.0", default-features = false, features = ["derive"] }
-scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false }
-sp-arithmetic = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false }
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false }
-
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false, optional = true }
-pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false, optional = true }
-pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17",  default-features = false, optional = true }
-frame-benchmarking = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.17', default-features = false, optional = true }
-serde = { version = "1.0.136", optional = true }
-
-[features]
-default = ["std"]
-std = [
-	"codec/std",
-	"sp-core/std",
-	"scale-info/std",
-	"sp-std/std",
-	"serde/std",
-	"frame-support/std",
-	"frame-system/std",
-	"pallet-timestamp/std",
-	"pallet-balances/std",
-]
-runtime-benchmarks = [
-	"frame-benchmarking",
-	"frame-support/runtime-benchmarks",
-	"frame-system/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-]
-try-runtime = ["frame-support/try-runtime"]
